VapourSynth.LWLibavSource: add bint soft_reset = 1 parameter
I think the new soft reset behavior is fairly safe: tested random
seeking on mpeg2, avc and hevc files. And the user can always switch
it off if the new reset behavior turns out to be problematic.

Also tested the case in FFMS/ffms2#394 and confirmed that lsmas is
still unaffected.

Fixes #10.

Signed-off-by: akarin <i@akarin.info>