common/lwindex.c: zero-initialize all 'video_info' and 'audio_info' elements
The `video_info` and `audio_info` arrays used to be created with `lw_malloc_zero` (which performs zero-initialization) but resized with `realloc` (which doesn't zero-initialize the newly allocated elements). When registering a new video or audio sample, the array is considered to contain one more element but not all of its fields are assigned a value. So, to avoid accessing uninitialized fields later on, zero-initialize the whole element every time a new sample is registered.

Use `malloc` instead of `lw_malloc_zero` to make it clear that the initial array is not expected to be zero-initialized.